chore(refactor): Restructure project with comprehensive documentation and data generation scripts
