--- Client Server Options
@@ -1,18 +1,0 @@
-In the chat shell, various console commands may be used.
-
-Key commands: /show, /help, /set.
-/show gives a display of all changable options
-/help gives a list of useful tools
-/set changes ingame options
-
-"/set minplayers __" to adjust the number of players
-"/set timeout __" to adjust the time of each turn in seconds
-"/set endt __" to adjust the maximum number of turns
-
-To turn off map features, game/options/local client and uncheck every box that involves displaying something.
-
-In pregame lobby, right click on player and select "aitoggleplayer" to change player to AI
-
-To run "headless", set minplayers to 0, change to an AI, set timeout to 1 second, turn off map features. Optionally put max turns.
-
-Also possible to run just the server without client. Can use console commands to control, however it is difficult to find final scores without client.
