--- Scenarios
@@ -1,19 +1,0 @@
-1: British Isles - Medium size. 85x80. Max players = 5. End turn = 410. All victories enabled. Tech level = 3.
-
-2: Earth - Large. 160x90. Max players = 39. End turn = default (5000). All victories enabled. Tech level = 0.
-
-3: Earth - Medium. 120x60. Max players = 30. End turn = default (5000). All victories enabled. Tech level = 3.
-
-4: Earth - Small. 80x50. Max players = 39. End turn default (5000). All victories enabled. Tech level = 0.
-
-5: Earth - Huge. 200x100. Max players = 37. End turn = default (5000). All victories enabled. Tech level = 3.
-
-6: France - Large. 140x90. Max players = 26. End turn = default (5000). All victories enabled. Tech level = 2.
-
-7: Iberia - Large. 136x100. Max players = 8. End turn = default (5000). All victories enabled. Tech level = 3.
-
-8: Italy - Medium. 100x100. Max players = 9. End turn = default (5000). All victories enabled. Tech level = 3.
-
-9: Japan - Medium. 88x100. Max players = 8. End turn = default (5000). All victories enabled. Tech level = 3.
-
-10: North America - Medium. 116x100. Max players = 12. End turn = default (5000). All victories enabled. Tech level = 3.
