+++ docs/scenarios.md
@@ -1,0 +1,93 @@
+# FreeCiv Scenarios
+
+## Overview
+
+This document outlines the predefined scenarios available for use in FreeCiv, each with unique settings and configurations.
+
+## Scenario Details
+
+### 1. British Isles
+
+- **Map Size:** Medium (85x80)
+- **Max Players:** 5
+- **End Turn:** 410
+- **Victory Conditions:** All enabled
+- **Tech Level:** 3
+
+### 2. Earth (Large)
+
+- **Map Size:** Large (160x90)
+- **Max Players:** 39
+- **End Turn:** Default (5000)
+- **Victory Conditions:** All enabled
+- **Tech Level:** 0
+
+### 3. Earth (Medium)
+
+- **Map Size:** Medium (120x60)
+- **Max Players:** 30
+- **End Turn:** Default (5000)
+- **Victory Conditions:** All enabled
+- **Tech Level:** 3
+
+### 4. Earth (Small)
+
+- **Map Size:** Small (80x50)
+- **Max Players:** 39
+- **End Turn:** Default (5000)
+- **Victory Conditions:** All enabled
+- **Tech Level:** 0
+
+### 5. Earth (Huge)
+
+- **Map Size:** Huge (200x100)
+- **Max Players:** 37
+- **End Turn:** Default (5000)
+- **Victory Conditions:** All enabled
+- **Tech Level:** 3
+
+### 6. France (Large)
+
+- **Map Size:** Large (140x90)
+- **Max Players:** 26
+- **End Turn:** Default (5000)
+- **Victory Conditions:** All enabled
+- **Tech Level:** 2
+
+### 7. Iberia (Large)
+
+- **Map Size:** Large (136x100)
+- **Max Players:** 8
+- **End Turn:** Default (5000)
+- **Victory Conditions:** All enabled
+- **Tech Level:** 3
+
+### 8. Italy (Medium)
+
+- **Map Size:** Medium (100x100)
+- **Max Players:** 9
+- **End Turn:** Default (5000)
+- **Victory Conditions:** All enabled
+- **Tech Level:** 3
+
+### 9. Japan (Medium)
+
+- **Map Size:** Medium (88x100)
+- **Max Players:** 8
+- **End Turn:** Default (5000)
+- **Victory Conditions:** All enabled
+- **Tech Level:** 3
+
+### 10. North America (Medium)
+
+- **Map Size:** Medium (116x100)
+- **Max Players:** 12
+- **End Turn:** Default (5000)
+- **Victory Conditions:** All enabled
+- **Tech Level:** 3
+
+## Scenario Use Cases
+
+1. **Testing AI Strategies:** Utilize different map sizes and player counts to evaluate AI performance.
+2. **Balanced Gameplay:** Select scenarios with varying tech levels for diverse gameplay experiences.
+3. **Historical Reenactments:** Choose region-specific scenarios for more immersive gameplay.
