+++ docs/useful-functions.md
@@ -1,0 +1,125 @@
+# Useful Functions in FreeCiv for AI Data Extraction
+
+## Overview
+
+This document provides detailed documentation of key functions from FreeCiv that are essential for extracting game data used in the AI system. These functions are crucial for analyzing player performance, tracking game state, and informing strategic decisions within the AI.
+
+## Data Utilization in AI
+
+The data extracted from these functions is used for:
+
+- **Player Performance Analysis:** Tracking scores, gold, and technological advancements.
+- **Strategic Decision-Making:** Informing AI decisions based on current government and city development.
+- **Game State Assessment:** Providing comprehensive data for neural network training and analysis.
+
+## Function Reference
+
+### 1. `player_name(pplayer)`
+
+**Description:**  
+This function retrieves the name of a specified player. It takes a pointer to a player object as input and returns the player's name as a string.
+
+**Example:**
+
+```python
+player_name(pplayer) -> "Player1"
+```
+
+**Use Case:**  
+Used to identify and track individual players in the game for AI analysis.
+
+---
+
+### 2. `get_civ_score(pplayer)`
+
+**Description:**  
+Returns the current civilization score of the specified player. The score is a comprehensive metric representing the player's progress and achievements in the game.
+
+**Example:**
+
+```python
+get_civ_score(pplayer) -> 1500
+```
+
+**Use Case:**  
+Used to evaluate overall player performance and progress over time.
+
+---
+
+### 3. `get_gold(pplayer)`
+
+**Description:**  
+Provides the total amount of gold that the specified player currently possesses. Gold is a crucial resource in FreeCiv for various in-game actions.
+
+**Example:**
+
+```python
+get_gold(pplayer) -> 500
+```
+
+**Use Case:**  
+Monitors the player's economic status and resource management.
+
+---
+
+### 4. `get_gov(pplayer)`
+
+**Description:**  
+Returns the current government type of the specified player. The government type affects various game mechanics and bonuses.
+
+**Example:**
+
+```python
+get_gov(pplayer) -> "Republic"
+```
+
+**Use Case:**  
+Analyzes the player's current governance model and its implications on gameplay.
+
+---
+
+### 5. `get_techs(pplayer)`
+
+**Description:**  
+Returns the number of technologies the specified player has researched. This function is used to determine the player's technological advancement.
+
+**Example:**
+
+```python
+get_techs(pplayer) -> 15
+```
+
+**Use Case:**  
+Evaluates the player's technological progress and research priorities.
+
+---
+
+### 6. `get_cities(pplayer)`
+
+**Description:**  
+Provides the number of cities that the specified player controls. Cities are fundamental to resource production and growth.
+
+**Example:**
+
+```python
+get_cities(pplayer) -> 8
+```
+
+**Use Case:**  
+Monitors urban development and expansion strategies.
+
+---
+
+## File References
+
+Trait logic is implemented in:
+
+- `\ai\aicity` (City management)
+- `\ai\advdiplomacy` (Diplomatic decisions)
+- `\ai\advdomestic` (Economic planning)
+
+## AI Development Use Cases
+
+1. **Scenario Design:** Create custom AI personalities by combining trait values
+2. **Balance Testing:** Compare different trait configurations
+3. **Adaptive AI:** Implement dynamic trait adjustments based on game state
