+++ requirements.txt
@@ -1,0 +1,2 @@
+tensorflow==2.11.0
+numpy==1.23.4
